Moved GetRepositories tests to gormdb_repositories_test.go
